refactor: code hygiene — type alias, module scoping, naming conventions
- Alias InlineFindingEventPayload = AgentFindingRecord in findings.ts,
  eliminating a structurally identical duplicate type
- Move estimateTokens out of src/utils/ (top-level bucket) into
  src/agent/utils.ts; inline the one-liner directly in boundaries/
  user-instruction-loader.ts to keep cross-module imports clean;
  delete src/utils/
- Trim README agent mode section: remove internal implementation
  details (rule unit precomputation, capability tier fallback chains)
  that are not user-facing
- Add clarifying comment to makePromptVariant test factory
- Add three coding convention rules to CLAUDE.md: no top-level
  src/utils/, centralized prompt construction, self-documenting names
